refactor(service): split board/column DTOs under 300 lines and dedupe create helper
Extract the create-path Option->FieldUpdate mapping into shared conv.rs; split board
DTOs into requests/conversions; promote columns to a directory module. Guard the Patch
NoChange serialize-omission; document the no-upper-bound validation and response-Deserialize
decisions. All api files now under the 300-line guideline.
