+++ crates/kanban-service/src/api/v1/boards/mod.rs
@@ -1,3 +1,4 @@
+mod conversions;
 mod requests;
 mod response;
 pub use requests::{CreateBoardRequest, ReplaceBoardRequest, UpdateBoardRequest};
