+++ crates/kanban-service/src/api/v1/columns/mod.rs
@@ -1,0 +1,7 @@
+mod conversions;
+mod requests;
+mod response;
+pub use requests::{
+    CreateColumnRequest, ReorderColumnRequest, ReplaceColumnRequest, UpdateColumnRequest,
+};
+pub use response::ColumnResponse;
