+++ crates/kanban-service/src/api/v1/mod.rs
@@ -1,5 +1,6 @@
 mod boards;
 mod columns;
+mod conv;
 mod enums;
 mod error;
 mod error_mapping;
